Files and Folders refactor
