+++ .gitignore
@@ -4,17 +4,19 @@
 
 knowledge/
 *.pdf
+*.gguf
 
 # -- Rust ------------------------------------------------------------------- #
 # -- ---- ------------------------------------------------------------------- #
 
 /debug
-/target
-Cargo.lock
+/rust/target
+/rust/Cargo.lock
 *.gguf
 
 # -- Python ----------------------------------------------------------------- #
 # -- ------ ----------------------------------------------------------------- #
 
-.env
+.env/
+*.cpython-311-darwin.so
