--- Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "molina"
-version = "0.0.1"
-edition = "2021"
-description = " Research Assistant "
-
-[dependencies]
-pyo3 = { version = "0.15", features = ["extension-module"] }
-llama_cpp_rs = { git = "https://github.com/tbogdala/rust-llama.cpp", version = "0.5.3" }
-lopdf = { version = "0.34.0", default-features = true }
-rayon = { version = "1.6"}
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-
