+++ README.md
@@ -1,12 +1,19 @@
 # molina
 
-Welcome to the `molina` project, a synthetic research asistant for local knowledge representation.
+Welcome to the `molina` project, a synthetic research agent for local knowledge representation.
 
-Notice here the absence of the terms `Artificial`, `Intelliget` and similar, this is not an `artificially sweetened` project, is not a claim to be the panacea, neither the $n-th$ attempt to solve general let alone narrow intelligence, because this is definitely not intelligent. It is a tool, built with an agentic approach, for you interact with by the act of formulating research-related questions, which will get you responses using ONLY the academic papers you provide as the knowledge base.
+Notice here the absence of the terms `Artificial` and `Intelligence`, this is deliverate.
+This is not an `artificially sweetened` project, is not a claim to be the panacea, neither 
+the $n-th$ attempt to solve general let alone narrow intelligence because this is definitely
+not a claim of something being intelligent. 
+
+What is this then ?, a tool, built with an agentic approach, for you to interact with by the 
+act of formulating research-related questions, which will get you responses using *only* the 
+academic papers you provide as the *knowledge base*.
 
 The name is in honor to THE greatest Mexican researcher of all times, [Mario Molina (1943-2020)](https://es.wikipedia.org/wiki/Mario_Molina_(químico)).
 
-## Problem
+## Problems
 
 Challenges of the use of Large Language Models (LLMs), as an academic research assistant, are:  
 
@@ -15,4 +22,10 @@ Challenges of the use of Large Language Models (LLMs), as an academic research a
 - Outdated information.
 - Bias in Training Data.
 - Interpretability and Response Attribution.
+
+## Project Structure
+
+- Python
+- Rust
+- Models
