+++ python/molina/main.py
@@ -1,8 +1,6 @@
 from molina import read_pdf
 
 def main():
-    # result = add(5, 7)
-    # print(f"The result of addition is now: {result}")
     read_pdf("knowledge/case_1/Advances-Prospect-Kahneman-Tversky-1992.pdf")
 
 if __name__ == "__main__":
