+++ rust/src/content/README.md
@@ -1,0 +1,1 @@
+ ## Content Module
