+++ rust/src/data/README.md
@@ -1,0 +1,1 @@
+## Data Module
