+++ rust/src/data/db.rs
@@ -1,0 +1,2 @@
+// Placeholder
+
