+++ rust/src/data/mod.rs
@@ -1,0 +1,1 @@
+// Placeholder
