+++ rust/src/inference/README.md
@@ -1,0 +1,1 @@
+## Inference Module
