+++ rust/src/inference/mod.rs
@@ -1,0 +1,1 @@
+pub mod llama;
