+++ rust/src/messages/errors.rs
@@ -1,0 +1,2 @@
+
+// Placeholder
