+++ rust/src/messages/logs.rs
@@ -1,0 +1,1 @@
+// Placeholder
